+++ ckanext/dalrrd_emc_dcpr/public/base/less/main.less
@@ -1,5 +1,7 @@
 @ckan-path: "/home/appuser/ckan-frontend/less/";
 
+@import "@{ckan-path}bootstrap.less";
+
 @import "@{ckan-path}variables.less";
 
 // put our own variables here
