+++ ckanext/dalrrd_emc_dcpr/public/base/less/variables.less
@@ -1,2 +1,4 @@
 // temporary test
 @layoutTextColor: yellow;
+
+@mastheadTextColor: #ff0;
