+++ package.json
@@ -1,20 +1,13 @@
 {
-  "name": "ckan",
-  "version": "1.0.0",
-  "description": "CKAN: The Open Source Data Portal Software ==========================================",
+  "name": "ckanext-dalrrd-emc-dcpr",
+  "version": "0.1.0",
+  "description": "SASDI's EMC and DCPR ==========================================",
   "dependencies": {
-    "blueimp-file-upload": "^10.8.0",
-    "bootstrap": "^3.4.1",
-    "font-awesome": "^4.7.0",
     "gulp": "^4.0.2",
     "gulp-if": "^3.0.0",
     "gulp-less": "^4.0.1",
     "gulp-rename": "^2.0.0",
-    "gulp-sourcemaps": "^2.6.5",
-    "jquery": "^3.5.0",
-    "less": "^3.11.2",
-    "moment": "^2.24.0",
-    "qs": "^6.9.4"
+    "gulp-sourcemaps": "^2.6.5"
   },
   "devDependencies": {
     "cypress": "^4.8.0",
@@ -22,19 +15,18 @@
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "watch": "gulp watch",
-    "build": "gulp build",
-    "postinstall": "gulp updateVendorLibs"
+    "docker-watch": "gulp --gulpfile gulpfile-docker.js watch",
+    "docker-build": "gulp --gulpfile gulpfile-docker.js build"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/ckan/ckan.git"
+    "url": "git+https://github.com/kartoza/ckanext-dalrrd-emc-dcpr.git"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/ckan/ckan/issues"
+    "url": "https://github.com/kartoza/ckanext-dalrrd-emc-dcpr/issues"
   },
-  "homepage": "https://github.com/ckan/ckan#readme"
+  "homepage": "https://github.com/kartoza/ckanext-dalrrd-emc-dcpr#readme"
 }
